Implemented display board inheritance and updated board for 2d representation
