+++ autoplay/BoardReplay.py
@@ -1,42 +1,21 @@
 import tkinter as tk
+from game.DisplayBase import DisplayBase, BTN_FONT, LBL_FONT, VAL_FONT
 
 from game.Direction import Direction
 from game.BoardState import BoardState
 
-CELL_COLORS = ["#9e948a", "#eee4da", "#ede0c8", "#f2b179", "#f59563", "#f67c5f",\
-               "#f65e3b", "#edcf72", "#edcc61", "#edc850", "#edc53f", "#edc22e"]
-TEXT_COLORS = ["#9e948a", "#776e65", "#776e65", "#f9f6f2", "#f9f6f2", "#f9f6f2",\
-               "#f9f6f2", "#f9f6f2", "#f9f6f2", "#f9f6f2", "#f9f6f2", "#f9f6f2"]
-FONT = ("Verdana", 40, "bold")
-BACKGROUND_COLOR = "#92877d"
-TITLE = "DK Holdings Presents: 2048"
-SIZE_PX = 500
-GRID_SPACING = 10 # space between grid squares
-TXT_WIDTH = 4
-TXT_HEIGHT = 2
-
-BTN_FONT = ("Helvetica", 24, "bold")
-LBL_FONT = ("Helvetica", 16, "bold")
-VAL_FONT = ("Helvetica", 16, "")
-
-class BoardReplay(tk.Tk): # inherit from the tkinter frame object
+class BoardReplay(DisplayBase): # inherit from the tkinter frame object
     def __init__(self, board_state: BoardState):
-        # create board state
-        self.board = board_state
         self.current_move = 0
+        self.show_end = False # determines whether the end state of a move should be displayed instead of the start state
+        DisplayBase.__init__(self, board_state)        
 
-        # initialize window and frame
-        tk.Tk.__init__(self)
-        self.title(TITLE)
-        self.frame = tk.Frame(self)
-        self.frame.pack(side=tk.TOP)
-
+    def key_binds(self):
         # bind all keys
         self.bind("<Left>", self.key_down)
         self.bind("<Right>", self.key_down)
-        self.bind("x", self.quit_program) # press 'x' to quit
-        self.bind("q", self.quit_program) # press 'x' to quit   
 
+    def custom_components(self):
         # initialize control buttons
         self.btn_to_beginning = tk.Button(self, text="<<", font=BTN_FONT, command=self.go_beginning)
         self.btn_to_beginning.pack(side=tk.LEFT)
@@ -55,90 +34,43 @@ class BoardReplay(tk.Tk): # inherit from the tkinter frame object
         self.lbl_next_move = tk.Label(self, text="", font=VAL_FONT)
         self.lbl_next_move.pack(side=tk.LEFT)
 
-        # initialize grid
-        self.grid_cells = []
-        self.init_grid()
-        self.update_board()
-        self.update_grid()
-        self.mainloop()
-
     def go_forward(self):
         last_move_index = len(self.board.move_history) - 1
         already_at_end = self.current_move >= last_move_index
         if not already_at_end:
             self.current_move += 1
-            self.update_board()
         else:
             self.set_end()
         self.update_grid()
 
     def go_back(self):
+        self.btn_to_end.config(state=tk.NORMAL)
+        self.btn_forward.config(state=tk.NORMAL)
         if self.current_move > 0:
             self.current_move -= 1
-            self.update_board()
-            self.update_grid()
+        self.update_grid()
 
     def go_beginning(self):
-        if self.current_move > 0:
-            self.current_move = 0
-            self.update_board()
-            self.update_grid()
+        self.current_move = 0
+        self.go_back()
 
     def go_end(self):
         last_move_index = len(self.board.move_history) - 1
         self.current_move = last_move_index
-        self.update_board()
         self.set_end()
         self.update_grid()
 
     def update_board(self):
         mv = self.board.move_history[self.current_move]
         self.lbl_move_num["text"] = str(self.current_move + 1)
-        self.board.Squares = mv.start_state
-        self.btn_to_end.config(state=tk.NORMAL)
-        self.btn_forward.config(state=tk.NORMAL)
+        self.board.field = mv.end_state if self.show_end else mv.start_state
         self.lbl_next_move.config(text=str(mv.direction))
 
     def set_end(self):
+        self.show_end = True
         self.lbl_next_move.config(text="(end)")
         self.btn_to_end.config(state=tk.DISABLED)
         self.btn_forward.config(state=tk.DISABLED)
-        self.board.Squares = self.board.move_history[self.current_move].end_state
-
-    def init_grid(self):
-        size = self.board.BOARD_SIZE
-        background = tk.Frame(self.frame, bg=BACKGROUND_COLOR, \
-                                          width=size,          \
-                                          height=size)
-        background.grid()
-        for i in range(size):
-            grid_row = []
-            for j in range(size):
-                cell = tk.Frame(background, width=SIZE_PX/size, \
-                                           height=SIZE_PX/size)
-                cell.grid(row=i,                    \
-                          column=j,                 \
-                          padx=GRID_SPACING,        \
-                          pady=GRID_SPACING)
-                cell_text = tk.Label(master=cell,      \
-                                  justify=tk.CENTER,   \
-                                  font=FONT,           \
-                                  width=TXT_WIDTH,     \
-                                  height=TXT_HEIGHT)
-                cell_text.grid()
-                grid_row.append(cell_text)
-            self.grid_cells.append(grid_row)
-    
-    def update_grid(self):
-        for i in range(self.board.BOARD_SIZE):
-            for j in range(self.board.BOARD_SIZE):
-                val = self.board.get_value(j + 1, i + 1)
-                txt = self.board.get_display_value(j + 1, i + 1)
-                self.grid_cells[i][j].configure(text=txt,                             \
-                                                  bg=CELL_COLORS[((val+1) % 12) - 1], \
-                                                  fg=TEXT_COLORS[((val+1) % 12) - 1])
-     
-        self.update_idletasks() # performs rendering tasks while avoiding race conditions due to callbacks
 
     def key_down(self, e):
         if e.keysym == "Left":
@@ -146,6 +78,3 @@ class BoardReplay(tk.Tk): # inherit from the tkinter frame object
         if e.keysym == "Right":
             if self.current_move < (len(self.board.move_history) - 1):
                 self.go_forward()
-
-    def quit_program(self, e):
-        quit()
